+++ aerende/configuration.py
@@ -1,8 +1,35 @@
+# -*- coding: utf-8 -*-
+
+"""
+aerende.configuration
+------------------
+
+This module contains the configuration logic for aerende. This includes the
+logic for loading a config file, writing a config file none is present, and
+the loading of default configuration options.
+
+There are 3 seperate configuration categories, at the moment. They are:
+
+palette   :: defines the colour palettes used various elements.
+data_options :: defines settings for aerende's data
+key_bindings :: defaults the key bindings for aerende
+"""
+
 from os import path, makedirs
 import yaml
 
 
 class Configuration(object):
+
+    # Default Configuration Options
+
+    # Default options for the palettes
+    # For more information on valid palette settings, see:
+    # http://urwid.org/manual/displayattributes.html#standard-foreground-colors
+    # status_bar :: the palette for the status bar at the bottom
+    # edit_bar :: the palette for the note editing bar
+    # highlight_note :: the palette for the currently focused note
+    # high_priority :: the palette for any high priority (>= 10) note
 
     DEFAULT_PALETTE = {
         'status_bar': ['black', 'white'],
@@ -11,9 +38,22 @@ class Configuration(object):
         'high_priority': ['light red', 'default']
     }
 
-    DEFAULT_DATA_PATH = {
+    # Default options for the data options.
+    # data_path :: path to the aerende YAML data file
+
+    DEFAULT_DATA_OPTIONS = {
         'data_path': '~/.andgeloman/aerende/data.yml'
     }
+
+    # Default key bindings
+    # new_note :: key to create a new note
+    # delete_note :: key to delete the focused note
+    # edit_note :: key to edit the focused note
+    # increment_note_priority :: key to increment the focused note's priority
+    # decrement_note_priority :: key to decrement the focused note's priority
+    # quit :: key to exit aerende
+    # next_note :: focus the next note
+    # previous_note :: focus the previous note
 
     DEFAULT_KEY_BINDINGS = {
         'new_note': 'n',
@@ -28,14 +68,23 @@ class Configuration(object):
 
     DEFAULT_CONFIG = {
         'palette': DEFAULT_PALETTE,
-        'data_path': DEFAULT_DATA_PATH,
+        'data_options': DEFAULT_DATA_OPTIONS,
         'key_bindings': DEFAULT_KEY_BINDINGS
     }
 
     def __init__(self, configuration_path):
+        """ On initialisation, preload the configuration options from the
+        defaults.
+        """
+        self.palette = self.DEFAULT_PALETTE
+        self.data_path = self.DEFAULT_DATA_OPTIONS
+        self.key_bindings = self.DEFAULT_KEY_BINDINGS
         self.__load_configuration(configuration_path)
 
     def __load_configuration(self, configuration_path):
+        """ Load the configuration from the supplied path. If the file does
+        not exist at this path, create it from the default config settings.
+        """
         expanded_path = path.expanduser(configuration_path)
         if not path.exists(expanded_path):
             makedirs(path.dirname(expanded_path))
@@ -43,40 +92,32 @@ class Configuration(object):
                 yaml.dump(self.DEFAULT_CONFIG, config_file,
                           default_flow_style=False)
             self.palette = self.DEFAULT_PALETTE
-            self.data_path = self.DEFAULT_DATA_PATH
+            self.data_path = self.DEFAULT_DATA_OPTIONS
             self.key_bindings = self.DEFAULT_KEY_BINDINGS
         else:
             self.__load_configuration_values(expanded_path)
 
     def __load_configuration_values(self, path):
+        """ Load the configuration file, update the config values from this
+        file.
+        """
         with open(path, 'r') as config_file:
             config_dict = yaml.load(config_file)
-            if 'palette' in config_dict:
-                self.palette = self.__validate_configuration(
-                    'palette', config_dict, self.DEFAULT_PALETTE)
-            else:
-                self.palette = self.DEFAULT_PALETTE
 
-            if 'data_path' in config_dict:
-                self.data_path = self.__validate_configuration(
-                    'data_path', config_dict, self.DEFAULT_DATA_PATH)
-            else:
-                self.data_path = self.DEFAULT_DATA_PATH
+            config_variables = {
+                'palette': self.palette,
+                'data_options': self.data_path,
+                'key_bindings': self.key_bindings
+            }
 
-            if 'key_bindings' in config_dict:
-                self.key_bindings = self.__validate_configuration(
-                    'key_bindings', config_dict, self.DEFAULT_KEY_BINDINGS)
-            else:
-                self.key_bindings = self.DEFAULT_KEY_BINDINGS
+            for key, value in config_variables.items():
+                self.__update_configuration(key, config_dict, value)
 
-    def __validate_configuration(self, config_key, config_dict, defaults):
-        config_item = {}
-        for key, default_value in defaults.items():
-            if key in config_dict[config_key]:
-                config_item[key] = config_dict[config_key][key]
-            else:
-                config_item[key] = default_value
-        return config_item
+    def __update_configuration(self, config_key, config_dict, var):
+        """ Update a config dictionary given a category key
+        """
+        if config_key in config_dict:
+            var.update(config_dict[config_key])
 
     def get_palette(self):
         return [[k] + v for k, v in self.palette.items()]